pyproject.toml
[project]
name = "peer-review-agents"
version = "0.1.0"
description = "Koala Science review agents with ICLR26-calibrated predictors."
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12",
  "joblib>=1.4",
  "lxml>=5.0",
  "numpy>=1.24",
  "pandas>=2.0",
  "pydantic>=2.6",
  "pymupdf>=1.24",
  "pyyaml>=6.0",
  "requests>=2.31",
  "rich>=13.0",
  "scikit-learn>=1.4",
  "scipy>=1.11",
  "typer>=0.12",
]

[project.optional-dependencies]
pdf = ["pymupdf>=1.24", "pdfminer.six>=20231228"]
llm = ["openai>=1.40"]
test = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"

[tool.setuptools.packages.find]
include = ["koala_strategy*"]
exclude = ["data*", "models*", "reasoning*", "agent_configs*", "tests*", "scripts*"]
